Start testing and finish up some packet stuff #7

+++ Protocol/src/main/java/gdx/lunar/protocol/packet/Packet.java
@@ -62,7 +62,7 @@ public abstract class Packet {
      * Write PID
      */
     protected void writeId() {
-        buffer.writeByte(getId());
+        buffer.writeInt(getId());
     }
 
     /**
